+++ apps/backoffice-v2/src/common/components/templates/Providers/Providers.tsx
@@ -3,11 +3,14 @@ import { FunctionComponent, PropsWithChildren } from 'react';
 
 import { queryClient } from '@/lib/react-query/query-client';
 import { AuthProvider } from '@/domains/auth/context/AuthProvider/AuthProvider';
+import { TooltipProvider } from '@ballerine/ui';
 
 export const Providers: FunctionComponent<PropsWithChildren> = ({ children }) => {
   return (
     <QueryClientProvider client={queryClient}>
-      <AuthProvider>{children}</AuthProvider>
+      <AuthProvider>
+        <TooltipProvider delayDuration={100}>{children}</TooltipProvider>
+      </AuthProvider>
     </QueryClientProvider>
   );
 };