+++ apps/backoffice-v2/src/domains/business-reports/query-keys.ts
@@ -17,26 +17,31 @@ export const businessReportsQueryKey = createQueryKeys('business-reports', {
     ...params
   }: {
     reportType?: MerchantReportType;
-    search: string;
-    page: number;
-    pageSize: number;
-    sortBy: string;
-    sortDir: string;
-    riskLevels: TRiskLevel[];
-    statuses: TReportStatusValue[];
-    findings: string[];
+    search?: string;
+    page?: number;
+    pageSize?: number;
+    sortBy?: string;
+    sortDir?: string;
+    riskLevels?: TRiskLevel[];
+    statuses?: TReportStatusValue[];
+    findings?: string[];
     from?: string;
     to?: string;
+    isAlert?: boolean;
   }) => ({
     queryKey: [{ page, pageSize, sortBy, sortDir, ...params }],
     queryFn: () => {
       const data = {
         ...params,
-        page: {
-          number: Number(page),
-          size: Number(pageSize),
-        },
-        orderBy: `${sortBy}:${sortDir}`,
+        ...(page && pageSize
+          ? {
+              page: {
+                number: Number(page),
+                size: Number(pageSize),
+              },
+            }
+          : {}),
+        ...(sortBy && sortDir ? { orderBy: `${sortBy}:${sortDir}` } : {}),
       };
 
       return fetchBusinessReports(data);