+++ packages/ui/src/components/templates/report/hooks/useReportTabs/useReportTabs.tsx
@@ -1,6 +1,7 @@
 import { Crown } from 'lucide-react';
 import { ComponentProps, ReactNode, useMemo } from 'react';
 import { Writable } from 'type-fest';
+import { ContentTooltip } from '@/components/molecules/ContentTooltip/ContentTooltip';
 
 import {
   AdsAndSocialMedia,
@@ -13,17 +14,21 @@ import {
   createReportAdapter,
 } from '@/components';
 
-export const useReportTabs = ({
-  reportVersion,
-  report,
-  companyName,
-  Link,
-}: {
+type UseReportTabsProps = {
   reportVersion: string;
+  isOnboarding: boolean;
   report: Record<PropertyKey, any>;
   companyName: string;
   Link: ComponentProps<typeof BusinessReportSummary>['Link'];
-}) => {
+};
+
+export const useReportTabs = ({
+  reportVersion,
+  isOnboarding,
+  report,
+  companyName,
+  Link,
+}: UseReportTabsProps) => {
   const adapter = createReportAdapter({
     reportVersion,
   });
@@ -64,9 +69,9 @@ export const useReportTabs = ({
       violations: websiteCredibilityAnalysis,
     },
     {
-      title: 'Ads and Social Media Analysis',
+      title: 'Social Media Analysis',
       search: '?activeTab=adsAndSocialMedia',
-      violations: adsAndSocialMediaAnalysis ?? [],
+      violations: null,
     },
     {
       title: 'Website Line of Business Analysis',
@@ -100,9 +105,28 @@ export const useReportTabs = ({
           value: 'summary',
           content: (
             <>
-              <h3 className={'mb-8 text-lg font-bold'}>Summary</h3>
+              <ContentTooltip
+                description={
+                  <p>
+                    Provides a concise overview of the merchant&apos;s risk level, integrating
+                    various factors into a clear summary for informed decisions.
+                  </p>
+                }
+                props={{
+                  tooltipContent: {
+                    className: 'max-w-[400px] whitespace-normal',
+                  },
+                  tooltipTrigger: {
+                    className: 'col-span-full text-lg font-bold',
+                  },
+                }}
+              >
+                <h3 className={'mb-8 text-lg font-bold'}>Summary</h3>
+              </ContentTooltip>
+
               <BusinessReportSummary
                 summary={summary}
+                isOnboarding={isOnboarding}
                 ongoingMonitoringSummary={ongoingMonitoringSummary}
                 riskScore={riskScore}
                 riskIndicators={riskIndicators as Writable<typeof riskIndicators>}
@@ -158,7 +182,7 @@ export const useReportTabs = ({
           ),
         },
         {
-          label: 'Ads and Social Media',
+          label: 'Social Media',
           value: 'adsAndSocialMedia',
           content: (
             <AdsAndSocialMedia
@@ -197,6 +221,7 @@ export const useReportTabs = ({
       formattedMcc,
       homepageScreenshotUrl,
       lineOfBusinessDescription,
+      isOnboarding,
       ongoingMonitoringSummary,
       onlineReputationAnalysis,
       pricingAnalysis,