+++ packages/eslint-config-react/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @ballerine/eslint-config-react
 
+## 2.0.31
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/eslint-config@1.1.31
+
+## 2.0.30
+
+### Patch Changes
+
+- version bump
+  s Please enter a summary for your changes.
+- Updated dependencies
+  - @ballerine/eslint-config@1.1.30
+
+## 2.0.29
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/eslint-config@1.1.29
+
 ## 2.0.28
 
 ### Patch Changes