+++ packages/react-pdf-toolkit/CHANGELOG.md
@@ -1,10 +1,49 @@
 # @ballerine/react-pdf-toolkit
 
-## 1.2.61
+## 1.2.66
 
 ### Patch Changes
 
 - version bump
+- Updated dependencies
+  - @ballerine/config@1.1.31
+  - @ballerine/ui@0.5.66
+
+## 1.2.65
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/ui@0.5.65
+
+## 1.2.64
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/ui@0.5.64
+
+## 1.2.63
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/ui@0.5.63
+
+## 1.2.62
+
+### Patch Changes
+
+- version bump
+  s Please enter a summary for your changes.
+- Updated dependencies
+  - @ballerine/config@1.1.30
+  - @ballerine/ui@0.5.62
+
+## 1.2.61
+
+### Patch Changes
+
 - Updated dependencies
   - @ballerine/ui@0.5.61
 
@@ -12,7 +51,9 @@
 
 ### Patch Changes
 
+- bump
 - Updated dependencies
+  - @ballerine/config@1.1.29
   - @ballerine/ui@0.5.60
 
 ## 1.2.59