+++ packages/workflow-core/CHANGELOG.md
@@ -1,10 +1,29 @@
 # @ballerine/workflow-core
 
-## 0.6.84
+## 0.6.86
 
 ### Patch Changes
 
 - version bump
+- Updated dependencies
+  - @ballerine/common@0.9.67
+
+## 0.6.85
+
+### Patch Changes
+
+- version bump
+  s Please enter a summary for your changes.
+- Updated dependencies
+  - @ballerine/common@0.9.66
+
+## 0.6.84
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.65
 
 ## 0.6.83