+++ sdks/web-ui-sdk/CHANGELOG.md
@@ -1,5 +1,30 @@
 # web-ui-sdk
 
+## 1.5.68
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.9.67
+
+## 1.5.67
+
+### Patch Changes
+
+- version bump
+  s Please enter a summary for your changes.
+- Updated dependencies
+  - @ballerine/common@0.9.66
+
+## 1.5.66
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.65
+
 ## 1.5.65
 
 ### Patch Changes