+++ sdks/workflow-node-sdk/CHANGELOG.md
@@ -1,9 +1,27 @@
 # @ballerine/workflow-node-sdk
 
+## 0.6.86
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.86
+
+## 0.6.85
+
+### Patch Changes
+
+- version bump
+  s Please enter a summary for your changes.
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.85
+
 ## 0.6.84
 
 ### Patch Changes
 
+- bump
 - Updated dependencies
   - @ballerine/workflow-core@0.6.84