+++ services/workflows-service/CHANGELOG.md
@@ -1,11 +1,33 @@
 # @ballerine/workflows-service
 
-## 0.7.88
+## 0.7.90
 
 ### Patch Changes
 
 - version bump
 - Updated dependencies
+  - @ballerine/workflow-core@0.6.86
+  - @ballerine/workflow-node-sdk@0.6.86
+  - @ballerine/common@0.9.67
+
+## 0.7.89
+
+### Patch Changes
+
+- version bump
+  s Please enter a summary for your changes.
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.85
+  - @ballerine/workflow-node-sdk@0.6.85
+  - @ballerine/common@0.9.66
+
+## 0.7.88
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.65
   - @ballerine/workflow-core@0.6.84
   - @ballerine/workflow-node-sdk@0.6.84